+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
     - Add support to open forms in editors
     - Add an option to set threads limit to throw
+    - There is a new, much more powerfull global search system instead of the old per timeline search.
       
 ### Changes